(fix) Next cell isn't `undefined` under any circumstances
When ants where in "deliver food" mode, it was possible that they will
have no possible option to go back as returned by the `getHighestStep`
function. Bug was due to checking `prevPosition` x and y properties
separately (instead of together). Additionally, an extra clause to
go back from the same way it returned, if necessary, was added.