(feat) Ants "normalize" adjacent cells when stuck
When ants are surrounded by pheromones when delivering food, they will
get reduce the `steps` of the surrounding cells to the minimum `step`
of the cells surrounding this cells.

This is done to help ants find a path when they enter a loop where
the highest step cells are the same.